% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_schema.R
\name{create_schema}
\alias{create_schema}
\title{기본 함수를 가져 옴}
\usage{
create_schema(type_is = "s3", table_is)
}
\arguments{
\item{infile}{Path to the input file}
}
\value{
A mat
}
\description{
테스트 해볼 것
}
