% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_char_name.R
\name{get_char_name}
\alias{get_char_name}
\title{DataFrame에서 Date, TimeStamp 형식의 Columns 이름을 찾는 함수}
\usage{
get_char_name(df)
}
\arguments{
\item{infile}{Path to the input file}
}
\value{
A mat
}
\description{
테스트 해볼 것
}
