% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mysql_schema.R
\name{get_mysql_schema}
\alias{get_mysql_schema}
\title{기본 함수를 가져 옴}
\usage{
get_mysql_schema(df, table_is)
}
\arguments{
\item{infile}{Path to the input file}
}
\value{
A mat
}
\description{
테스트 해볼 것
}
