% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wotan_insert.R
\name{get_wotan_insert}
\alias{get_wotan_insert}
\title{기본 함수를 가져 옴}
\usage{
get_wotan_insert(df, target_name)
}
\arguments{
\item{infile}{Path to the input file}
}
\value{
A mat
}
\description{
테스트 해볼 것
}
