% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_loader.R
\name{gs_loader}
\alias{gs_loader}
\title{Google Sheet4를 사용해서 sheet 정보를 가져옴}
\usage{
gs_loader(sheet_is = "loader", project_is = "RRS")
}
\arguments{
\item{infile}{Path to the input file}
}
\value{
A matrix of the infile
}
\description{
테스트 해볼 것
}
