% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_update.R
\name{gs_update}
\alias{gs_update}
\title{Google Sheet4를 사용해서 sheet 정보를 가져옴}
\usage{
gs_update(df_is, sheet_is, project_is, is_write = F)
}
\arguments{
\item{infile}{Path to the input file}
}
\value{
A matrix of the infile
}
\description{
테스트 해볼 것
}
